requirements.txt
django==1.9.3
gunicorn==19.4.5
oauth2client==1.5.2
google-api-python-client==1.4.2
PyOpenSSL
futures==3.0.5
django-ipware==1.1.3